feat(domain/piece/): 도메인의 piece 내 클래스 패키지 구조화
