+++ src/main/java/chess/domain/piece/nonsliding/King.java
@@ -1,6 +1,9 @@
-package chess.domain.piece;
+package chess.domain.piece.nonsliding;
 
 import chess.domain.color.Color;
+import chess.domain.piece.Direction;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import java.util.Set;
 
 public class King extends NonSlidingPiece {
