+++ src/main/java/chess/domain/piece/nonsliding/NonSlidingPiece.java
@@ -1,6 +1,10 @@
-package chess.domain.piece;
+package chess.domain.piece.nonsliding;
 
 import chess.domain.color.Color;
+import chess.domain.piece.Direction;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import java.util.Set;
 
 public abstract class NonSlidingPiece extends Piece {
