+++ src/main/java/chess/domain/piece/pawn/BlackFirstPawn.java
@@ -1,12 +1,15 @@
-package chess.domain.piece;
+package chess.domain.piece.pawn;
 
 import chess.domain.color.Color;
+import chess.domain.piece.Direction;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import java.util.Set;
 
-public class BlackPawn extends Pawn {
-    private static Set<Direction> DIRECTIONS = Set.of(Direction.DOWN, Direction.LEFT_DOWN, Direction.RIGHT_DOWN);
+public class BlackFirstPawn extends Pawn {
+    private static Set<Direction> DIRECTIONS = Direction.getBlackFirstPawnDirection();
 
-    public BlackPawn(Position position) {
+    public BlackFirstPawn(Position position) {
         super(position, Color.BLACK, DIRECTIONS);
     }
