+++ src/main/java/chess/domain/piece/pawn/WhiteFirstPawn.java
@@ -1,6 +1,9 @@
-package chess.domain.piece;
+package chess.domain.piece.pawn;
 
 import chess.domain.color.Color;
+import chess.domain.piece.Direction;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import java.util.Set;
 
 public class WhiteFirstPawn extends Pawn {
