+++ src/main/java/chess/domain/piece/sliding/SlidingPiece.java
@@ -1,6 +1,10 @@
-package chess.domain.piece;
+package chess.domain.piece.sliding;
 
 import chess.domain.color.Color;
+import chess.domain.piece.Direction;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
 import java.util.Set;
 
 public abstract class SlidingPiece extends Piece {
