+++ src/main/java/chess/domain/strategy/MoveStrategy.java
@@ -1,6 +1,6 @@
 package chess.domain.strategy;
 
-import chess.domain.piece.Blank;
+import chess.domain.piece.blank.Blank;
 import chess.domain.color.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
