+++ src/test/java/chess/domain/ChessGameTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 
 import chess.domain.color.Color;
 import chess.domain.piece.Position;
-import chess.domain.piece.Rook;
+import chess.domain.piece.sliding.Rook;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
