+++ src/test/java/chess/domain/piece/nonsliding/KnightTest.java
@@ -1,9 +1,12 @@
-package chess.domain.piece;
+package chess.domain.piece.nonsliding;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import chess.domain.color.Color;
+import chess.domain.piece.PieceType;
+import chess.domain.piece.Position;
+import chess.domain.piece.nonsliding.Knight;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
