+++ src/test/java/chess/domain/strategy/GeneralMoveStrategyTest.java
@@ -3,15 +3,15 @@ package chess.domain.strategy;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-import chess.domain.piece.Bishop;
+import chess.domain.piece.sliding.Bishop;
 import chess.domain.color.Color;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
+import chess.domain.piece.nonsliding.King;
+import chess.domain.piece.nonsliding.Knight;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.domain.piece.Position;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.sliding.Queen;
+import chess.domain.piece.sliding.Rook;
 import chess.domain.TestBoardFactory;
 import java.util.Map;
 import org.junit.jupiter.api.Assertions;
